requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
cycler==0.11.0
fonttools==4.36.0
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
metalog==0.2.2
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywatchman==1.4.1
requests==2.28.1
scipy==1.9.0
six==1.16.0
urllib3==1.26.12
Flask==1.1.2
Werkzeug==1.0.1
Jinja2==2.11.2
jsonify==0.5
MarkupSafe==1.1.1
itsdangerous==1.1.0
click==7.1.2
chardet==3.0.4
pycoingecko==2.1.0
schedule==1.1.0
python-dotenv==0.21.0
boto3==1.17.103